refactor(scripts): share the workspace walk, not just the refusal it contains
/simplify, four angles. Two agents independently found the same structural
miss: extracting existsOrThrow shared 15 lines and left 32 duplicated, because
listPackages then became byte-identical in check-test-wiring and
check-test-glob-coverage - differing only in a constant's name, 'utf-8' vs
'utf8', and the wording of a 12-line comment that was itself duplicated. The
first commit's own docstring argues that two walks kept in agreement by whoever
edited one remembering the other "is not a property, it is a habit, and #3347 is
what it costs". That argument is exactly as true across two files.

scripts/lib/list-workspace-packages.mjs now owns the walk and the parents
constant. 58 lines out of one gate, 32 out of the other.

It also states the invariant the altitude pass distilled, which is narrower and
more useful than "be loud everywhere":

    unreadable must never shrink a population a gate reports a count of.

That explains why a refusal and a calibrated floor are alternatives rather than
a hierarchy, and why verify-npm-publish and rust-major-offset are right to
warn-and-floor instead.

NOT migrated, and the reason is load-bearing: scripts/docs/check-package-readmes.mjs
keeps its own copy. Two reviewers ranked folding it in as the top finding, both
citing that the repo already documents it as a duplicate. Neither checked why.
Its regression harness copyFileSync's THAT ONE FILE into a synthetic tree, so it
must stay import-free beyond node builtins; migrating it turned all 9 of its
tests red while the gate still passed against the real repo, because the failure
only exists in the synthetic tree where ../lib/ does not. The constraint is now
written in the file.

Three claims of mine that were false:

- "A PROJECTION of listPackages, not a second walk" was true of the source and
  false of the run: it is called once per reader, so there are still two walks.
  Measured at 0.96 ms of a 100 ms gate, which is why it stays two.
- "#3350 fixed exactly this in two sibling gates" - it fixed three.
- samePath's realpath half was justified by a symlinked-checkout divergence.
  Measured on four invocations - through a symlink, with --preserve-symlinks,
  invoking through the link, and both - and path.resolve matched every time,
  because getcwd(3) returns the resolved path and Node's ESM loader realpaths
  the module URL. Kept as defence for unmeasured platforms, but it no longer
  claims to catch a case that occurs here.

check-ci-path-coverage then caught a consequence I had not thought about: moving
the apps/ walk into the lib made the allowlist row exempting check-test-wiring
from apps/landing stale. Deleted, per that gate's own advice that a stale
exemption hides the next hole. 36 reasoned exemptions now, down from 37.

1056 script tests pass, lint clean, all four gates green on the real repo.
