+++ scripts/typecheck-tests.mjs
@@ -252,12 +252,17 @@ const USAGE_LINES = [
 ];
 
 /**
- * Do `a` and `b` name the same directory?
+ * Are these the same directory?
  *
- * `process.cwd()` reports the REALPATH while REPO_ROOT is derived from this
- * file's own URL, so a symlinked checkout spells one directory two ways and a
- * string compare misses — and missing here admits exactly the run the caller
- * below refuses. Same reasoning as `rootReal` in scripts/check-changesets.mjs.
+ * `path.resolve` is the whole answer on every path this is actually reached
+ * through. The realpath fallback is defence for platforms this was not measured
+ * on, NOT a case that occurs here: four invocations were tried on macOS - cd
+ * through a symlink, the same with --preserve-symlinks, invoking through the
+ * link from the real directory, and that with --preserve-symlinks - and
+ * `resolve` matched in all four, because getcwd(3) already returns the resolved
+ * path and Node's ESM loader realpaths the module URL. The one live caller
+ * passes `process.cwd()`, so `pkgDir` is a realpath by construction. Costs one
+ * pair of syscalls, once per run.
  */
 function samePath(a, b) {
   if (path.resolve(a) === path.resolve(b)) return true;
